requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
python-vlc # LGPL-2.1+
#wpa_supplicant # Mozilla Public License, v. 2.0